# Implement schema squashing