--- a/schema/cassandra/cadence/versioned/s0.0-0.23/manifest.json
+++ b/schema/cassandra/cadence/versioned/s0.0-0.23/manifest.json
@@ -1,0 +1,8 @@
+{
+    "CurrVersion": "0.23",
+    "MinCompatibleVersion": "0.23",
+    "Description": "Squashed 0.0 -> 0.23 schema",
+    "SchemaUpdateCqlFiles": [
+      "schema_0_23.cql"
+    ]
+  }