--- a/schema/cassandra/cadence/versioned/s0.0-0.23/manifest.json
+++ b/schema/cassandra/cadence/versioned/s0.0-0.23/manifest.json
@@ -3,6 +3,7 @@
     "MinCompatibleVersion": "0.23",
     "Description": "Squashed 0.0 -> 0.23 schema",
     "SchemaUpdateCqlFiles": [
-      "schema_0_23.cql"
+      "schema_0_23.cql",
+      "data_0_23.cql"
     ]
   }