--- a/tools/cassandra/updateTask_test.go
+++ b/tools/cassandra/updateTask_test.go
@@ -22,6 +22,7 @@ package cassandra
 
 import (
 	"log"
+	"os"
 	"testing"
 
 	"github.com/stretchr/testify/suite"
@@ -71,4 +72,17 @@ func (s *UpdateSchemaTestSuite) TestVisibilityDryrun() {
 	defer client.Close()
 	dir := "../../schema/cassandra/visibility/versioned"
 	s.RunDryrunTest(buildCLIOptions(), client, "-k", dir, cassandra.VisibilityVersion)
+}
+
+func (s *UpdateSchemaTestSuite) TestShortcut() {
+	client, err := newTestCQLClient(s.DBName)
+	s.Nil(err)
+	defer client.Close()
+	dir := "../../schema/cassandra/cadence/versioned"
+
+	cqlshArgs := []string{"--cqlversion=3.4.4", "-e", "DESC KEYSPACE %s;"}
+	if cassandraHost := os.Getenv("CASSANDRA_HOST"); cassandraHost != "" {
+		cqlshArgs = append(cqlshArgs, cassandraHost)
+	}
+	s.RunShortcutTest(buildCLIOptions(), client, "-k", dir, "cqlsh", cqlshArgs...)
 }