**What changed?**

For incremental schema, it is now possible to take a shortcut from one version to another without applying incremental changes one by one.
The directory name should be in format `s<from_ver>-<to_ver>`, where

* ` <to_ver>` must be greater than `<from_ver>`
* `<to_ver>` must exist in incremental schema steps, i.e. a directory with valid manifest named `v_<to_ver>` must exist
* `<from_ver>` must either exist in incremental schema steps, or be equal to initial version `0.0`

Schema tool will automatically select the path with least steps. 

Squashed shortcut schema from version `0.0` to `0.23` included. This particular schema version is selected since it's on the tip of latest `0.11` Cadence release.

**Why?**

For Cassandra instances with large replication factors and latency schema statements might take up to a second for a trivial operation such as column creation. Given that current version on tip of master is `0.27`, the bootstrap time adds up and will grow with each schema update.

Applying goal-state schema would only create 30+ entities and would run up to 4x times faster on Cassandra setups above.

Applying `schema.cql` is an option for a new installation, but prevents further incremental upgrades.


**How did you test it?**

To measure speed improvements, application times for both schema and every incremental version is now logged. Tested by playing back incremental updates up to version `0.27` on docker instance

### Time improvements

#### Before the change

```
cadence_1      | 2020/05/11 20:35:15 Schema updated from 0.0 to 0.1, elapsed 2.247461928s
cadence_1      | 2020/05/11 20:35:17 Schema updated from 0.1 to 0.2, elapsed 1.631742425s
cadence_1      | 2020/05/11 20:35:18 Schema updated from 0.2 to 0.3, elapsed 836.043275ms
cadence_1      | 2020/05/11 20:35:19 Schema updated from 0.3 to 0.4, elapsed 633.861512ms
cadence_1      | 2020/05/11 20:35:19 Schema updated from 0.4 to 0.5, elapsed 394.238516ms
cadence_1      | 2020/05/11 20:35:20 Schema updated from 0.5 to 0.6, elapsed 1.130838727s
cadence_1      | 2020/05/11 20:35:24 Schema updated from 0.6 to 0.7, elapsed 3.86523542s
cadence_1      | 2020/05/11 20:35:24 Schema updated from 0.7 to 0.8, elapsed 303.648524ms
cadence_1      | 2020/05/11 20:35:25 Schema updated from 0.8 to 0.9, elapsed 457.736671ms
cadence_1      | 2020/05/11 20:35:25 Schema updated from 0.9 to 0.10, elapsed 436.88099ms
cadence_1      | 2020/05/11 20:35:28 Schema updated from 0.10 to 0.11, elapsed 2.453376415s
cadence_1      | 2020/05/11 20:35:31 Schema updated from 0.11 to 0.12, elapsed 3.061781143s
cadence_1      | 2020/05/11 20:35:32 Schema updated from 0.12 to 0.13, elapsed 1.82323551s
cadence_1      | 2020/05/11 20:35:33 Schema updated from 0.13 to 0.14, elapsed 336.098623ms
cadence_1      | 2020/05/11 20:35:34 Schema updated from 0.14 to 0.15, elapsed 799.870505ms
cadence_1      | 2020/05/11 20:35:34 Schema updated from 0.15 to 0.16, elapsed 192.859672ms
cadence_1      | 2020/05/11 20:35:34 Schema updated from 0.16 to 0.17, elapsed 378.112309ms
cadence_1      | 2020/05/11 20:35:35 Schema updated from 0.17 to 0.18, elapsed 387.9124ms
cadence_1      | 2020/05/11 20:35:36 Schema updated from 0.18 to 0.19, elapsed 1.129843878s
cadence_1      | 2020/05/11 20:35:36 Schema updated from 0.19 to 0.20, elapsed 195.834643ms
cadence_1      | 2020/05/11 20:35:36 Schema updated from 0.20 to 0.21, elapsed 178.099493ms
cadence_1      | 2020/05/11 20:35:37 Schema updated from 0.21 to 0.22, elapsed 902.453475ms
cadence_1      | 2020/05/11 20:35:37 Schema updated from 0.22 to 0.23, elapsed 531.753091ms
cadence_1      | 2020/05/11 20:35:38 Schema updated from 0.23 to 0.24, elapsed 181.113603ms
cadence_1      | 2020/05/11 20:35:38 Schema updated from 0.24 to 0.25, elapsed 160.556925ms
cadence_1      | 2020/05/11 20:35:38 Schema updated from 0.25 to 0.26, elapsed 531.831506ms
cadence_1      | 2020/05/11 20:35:38 Schema updated from 0.26 to 0.27, elapsed 110.09224ms
cadence_1      | 2020/05/11 20:35:38 All schema changes completed in 25.296508226s
```

##### After the change

```
cadence_1      | 2020/05/11 22:37:38 Schema updated from 0.0 to 0.23, elapsed 4.242590367s
cadence_1      | 2020/05/11 22:37:39 Schema updated from 0.23 to 0.24, elapsed 303.520763ms
cadence_1      | 2020/05/11 22:37:39 Schema updated from 0.24 to 0.25, elapsed 227.529843ms
cadence_1      | 2020/05/11 22:37:39 Schema updated from 0.25 to 0.26, elapsed 741.045226ms
cadence_1      | 2020/05/11 22:37:40 Schema updated from 0.26 to 0.27, elapsed 164.768217ms
cadence_1      | 2020/05/11 22:37:40 All schema changes completed in 5.679644717s
```

#### Data validitiy

##### Before the change

```
root@1370b14fd06a:/# cqlsh -e "DESC KEYSPACE cadence"|md5sum
bd705d28afcfe8b5a207a905af16f9e1  -
root@1370b14fd06a:/# cqlsh -k cadence --no-color -e "SELECT * FROM domains;"|md5sum
6672f72e06d45d845df8aa736825f11d  -
root@1370b14fd06a:/# cqlsh -k cadence --no-color -e "SELECT * FROM domains_by_name_v2;"|md5sum
69dcf3ee18d67a066915dd371e622ac3  -
```

##### After the change

```
root@21133ee99c79:/# cqlsh -e "DESC KEYSPACE cadence"|md5sum
bd705d28afcfe8b5a207a905af16f9e1  -

root@21133ee99c79:/# cqlsh -k cadence --no-color -e "SELECT * FROM domains;"|md5sum
6672f72e06d45d845df8aa736825f11d  -

root@21133ee99c79:/# cqlsh -k cadence --no-color -e "SELECT * FROM domains_by_name_v2;"|md5sum
69dcf3ee18d67a066915dd371e622ac3  -
```

**Potential risks**

As with any other schema change, there are risks involved. One of the concerns is mismatching schemas in `schema.cql` for `0.11` Cadence version (`v0.23` Cassandra schema) and playing back the incremental changes. Some are minor, such as column ordering, but there is a missing events table and mismatching data type for a column.

However, the tests above show that shortcut schema is identical to one produced by playing incremental changes back.