@emrahs 2020-05-12
@aleksejuber PR looks pretty good to me. I just left a minor comment. Besides that, have you considered automating your md5sum test in as an integration test (possibly in `tools/common/schema/test/updatetest.go`)? I think it would be good to have it so that we don't need to reinvent a manual way to verify each new squash in the future. It'd also prevent inconsistencies in case a contributor modifies these CQL files for some reason.