@aleksejuber 2020-05-13
> . Besides that, have you considered automating your md5sum test in as an integration test (possibly in `tools/common/schema/test/updatetest.go`)?

This is a very good point, the integration test would surely benefit here to ensure initial data/schema consistency. I believe the check can be done without modifying the Go code, using modified schema initialisation bash scripts, as in

1. List all the shortcut directories to determine the end version and rename/move them out
2. Apply versions as usual, but take snapshots after every end version
3. On by one move the shortcut directories back, remove keyspace, init version `0.0`, replay up to the end version, compare schemas, and remove the shortcut

This way we will run same commands as the schema init scripts and avoid missing something out in test setup.

In your opinion, should the integration tests be included in this PR or create a separate one for this issue?