@emrahs 2020-05-13
> In your opinion, should the integration tests be included in this PR or create a separate one for this issue?

I'm OK with either way. Regarding how we implement the tests, I'd prefer having those in `tools/updatetest.go` for consistency with other tests and to ensure they run together.